+++ src/info.py
@@ -1,4 +1,4 @@
 class Info:
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.21b2" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2024-08-20"
+    AppVersionStr = "0.22b1" # see https://peps.python.org/pep-0440/
+    AppDateStr = "2024-08-21"